refactor(api): move assessment routes and controller
