+++ api/lib/routes.js
@@ -1,9 +1,7 @@
 import * as accountRecovery from './application/account-recovery/index.js';
 import * as adminMembers from './application/admin-members/index.js';
 import * as activityAnswers from './application/activity-answers/index.js';
-import * as answers from './application/answers/index.js';
 import * as assessmentResults from './application/assessment-results/index.js';
-import * as assessments from './application/assessments/index.js';
 import * as authentication from './application/authentication/index.js';
 import * as authenticationOidc from './application/authentication/oidc/index.js';
 import * as badges from './application/badges/index.js';
@@ -55,9 +53,7 @@ const routes = [
   accountRecovery,
   adminMembers,
   activityAnswers,
-  answers,
   assessmentResults,
-  assessments,
   authentication,
   authenticationOidc,
   badges,
