+++ api/src/shared/application/assessments/assessment-controller.js
@@ -1,23 +1,23 @@
-import { DomainTransaction } from '../../infrastructure/DomainTransaction.js';
+import { DomainTransaction } from '../../../../lib/infrastructure/DomainTransaction.js';
 import { Serializer as JSONAPISerializer } from 'jsonapi-serializer';
-import { AssessmentEndedError } from '../../domain/errors.js';
-import { usecases } from '../../domain/usecases/index.js';
-import { usecases as devcompUsecases } from '../../../src/devcomp/domain/usecases/index.js';
-import * as events from '../../domain/events/index.js';
-import { logger } from '../../infrastructure/logger.js';
-import * as assessmentRepository from '../../../src/shared/infrastructure/repositories/assessment-repository.js';
-import * as assessmentSerializer from '../../infrastructure/serializers/jsonapi/assessment-serializer.js';
-import * as activitySerializer from '../../infrastructure/serializers/jsonapi/activity-serializer.js';
-import * as challengeSerializer from '../../infrastructure/serializers/jsonapi/challenge-serializer.js';
-import * as competenceEvaluationSerializer from '../../../src/evaluation/infrastructure/serializers/jsonapi/competence-evaluation-serializer.js';
+import { AssessmentEndedError } from '../../../../lib/domain/errors.js';
+import { usecases } from '../../../../lib/domain/usecases/index.js';
+import { usecases as devcompUsecases } from '../../../devcomp/domain/usecases/index.js';
+import * as events from '../../../../lib/domain/events/index.js';
+import { logger } from '../../../../lib/infrastructure/logger.js';
+import * as assessmentRepository from '../../infrastructure/repositories/assessment-repository.js';
+import * as assessmentSerializer from '../../../../lib/infrastructure/serializers/jsonapi/assessment-serializer.js';
+import * as activitySerializer from '../../../../lib/infrastructure/serializers/jsonapi/activity-serializer.js';
+import * as challengeSerializer from '../../../../lib/infrastructure/serializers/jsonapi/challenge-serializer.js';
+import * as competenceEvaluationSerializer from '../../../evaluation/infrastructure/serializers/jsonapi/competence-evaluation-serializer.js';
 import {
   extractLocaleFromRequest,
   extractUserIdFromRequest,
-} from '../../infrastructure/utils/request-response-utils.js';
-import * as certificationChallengeRepository from '../../infrastructure/repositories/certification-challenge-repository.js';
+} from '../../../../lib/infrastructure/utils/request-response-utils.js';
+import * as certificationChallengeRepository from '../../../../lib/infrastructure/repositories/certification-challenge-repository.js';
 
-import { Examiner } from '../../domain/models/Examiner.js';
-import { ValidatorAlwaysOK } from '../../domain/models/ValidatorAlwaysOK.js';
+import { Examiner } from '../../../../lib/domain/models/Examiner.js';
+import { ValidatorAlwaysOK } from '../../../../lib/domain/models/ValidatorAlwaysOK.js';
 
 const save = async function (request, h, dependencies = { assessmentRepository }) {
   const assessment = assessmentSerializer.deserialize(request.payload);
