+++ api/src/shared/application/assessments/index.js
@@ -1,9 +1,9 @@
 import Joi from 'joi';
-import { config } from '../../config.js';
+import { config } from '../../../../lib/config.js';
 import { assessmentController } from './assessment-controller.js';
-import { securityPreHandlers } from '../security-pre-handlers.js';
-import { assessmentAuthorization } from '../preHandlers/assessment-authorization.js';
-import { identifiersType } from '../../domain/types/identifiers-type.js';
+import { securityPreHandlers } from '../../../../lib/application/security-pre-handlers.js';
+import { assessmentAuthorization } from '../../../../lib/application/preHandlers/assessment-authorization.js';
+import { identifiersType } from '../../../../lib/domain/types/identifiers-type.js';
 
 const { featureToggles } = config;
