+++ api/src/shared/routes.js
@@ -1,0 +1,5 @@
+import * as assessmentsRoutes from './application/assessments/index.js';
+
+const sharedRoutes = [assessmentsRoutes];
+
+export { sharedRoutes };
