+++ api/tests/shared/acceptance/application/assessments/assessment-controller-get_test.js
@@ -1,10 +1,10 @@
-import { expect, generateValidRequestAuthorizationHeader, databaseBuilder } from '../../../test-helper.js';
-import { createServer } from '../../../../server.js';
-import { LOCALE } from '../../../../src/shared/domain/constants.js';
+import { expect, generateValidRequestAuthorizationHeader, databaseBuilder } from '../../../../test-helper.js';
+import { createServer } from '../../../../../server.js';
+import { LOCALE } from '../../../../../src/shared/domain/constants.js';
 
 const { FRENCH_SPOKEN } = LOCALE;
 
-import { Assessment } from '../../../../src/shared/domain/models/Assessment.js';
+import { Assessment } from '../../../../../src/shared/domain/models/Assessment.js';
 
 describe('Acceptance | API | assessment-controller-get', function () {
   let server;
