+++ api/tests/shared/acceptance/application/assessments/assessment-controller-post_test.js
@@ -1,6 +1,6 @@
-import { expect, generateValidRequestAuthorizationHeader, databaseBuilder } from '../../../test-helper.js';
-import { createServer } from '../../../../server.js';
-import { BookshelfAssessment } from '../../../../lib/infrastructure/orm-models/Assessment.js';
+import { expect, generateValidRequestAuthorizationHeader, databaseBuilder } from '../../../../test-helper.js';
+import { createServer } from '../../../../../server.js';
+import { BookshelfAssessment } from '../../../../../lib/infrastructure/orm-models/Assessment.js';
 
 describe('Acceptance | API | Assessments POST', function () {
   let server;
