+++ api/tests/shared/acceptance/application/assessments/assessment-controller-update-last-challenge-state_test.js
@@ -5,10 +5,10 @@ import {
   mockLearningContent,
   learningContentBuilder,
   knex,
-} from '../../../test-helper.js';
+} from '../../../../test-helper.js';
 
-import { createServer } from '../../../../server.js';
-import { Assessment } from '../../../../src/shared/domain/models/Assessment.js';
+import { createServer } from '../../../../../server.js';
+import { Assessment } from '../../../../../src/shared/domain/models/Assessment.js';
 
 const competenceId = 'recCompetence';
 const skillWeb2Id = 'recAcquisWeb2';
